project-spa-backend/python-api/requirements.txt
# project-spa-backend/python-api/requirements.txt
flask
python-docx
